Move base_fee::tests into a new file
